cpu/nrf5x: update GPIO_UNDEF value
The gpio_t value is uint8_t so use UINT8_MAX for GPIO_UNDEF

+++ cpu/nrf5x_common/include/periph_cpu_common.h
@@ -55,7 +55,7 @@ extern "C" {
 /**
  * @brief   Override GPIO_UNDEF value
  */
-#define GPIO_UNDEF          (UINT_MAX)
+#define GPIO_UNDEF          (UINT8_MAX)
 
 /**
  * @brief   Generate GPIO mode bitfields
